Refactor the encoder code
