+++ cbverifier/encoding/ts.py
@@ -1,0 +1,54 @@
+""" Data structure to represent a transition system
+
+"""
+
+from pysmt.shortcuts import TRUE as TRUE_PYSMT
+from pysmt.shortcuts import FALSE as FALSE_PYSMT
+from pysmt.shortcuts import And
+
+class TransitionSystem:
+    """ (symbolic) Transition system"""
+    def __init__(self):
+        # internal representation of the transition system
+        self.state_vars = set()
+        self.input_vars = set()
+        self.init = TRUE_PYSMT()
+        self.trans = TRUE_PYSMT()
+
+    def add_var(self, var):
+        self.state_vars.add(var)
+
+    def add_ivar(self, var):
+        self.input_vars.add(var)
+
+    def product(self, other_ts):
+        """ Computes the synchronous product of self with other_ts,
+        storing the product in self.
+
+        Given TS1 = <V1, W1, I1, T1> and TS2 = <V2, W2, I2, T2>
+        the product is the transition system
+        TSP = <V1 union V2, W1 union W2, I1 and I2, T1 and T2>
+
+        (V are the state variables, W the input variables,
+        I is the initial condition, T the transition relation)
+        """
+
+        self.state_vars.update(other_ts.state_vars)
+        self.input_vars.update(other_ts.input_vars)
+        self.init = And(self.init, other_ts.init)
+        self.trans = And(self.trans, other_ts.trans)
+
+    def __repr__(self):
+        """ Not efficient, need to use a buffer..."""
+
+        res = "State vars: "
+        for v in self.state_vars: res += ", %s" % v
+        res += "\nInput vars:"
+        for v in self.input_vars: res += ", %s" % v
+        res += "\nINIT: "
+
+        res += str(self.init.serialize())
+        res += "\nTRANS: "
+        res += str(self.trans.simplify().serialize())
+
+        return res
